+++ comics/comics/charliehorse.py
@@ -11,7 +11,7 @@ class Meta(MetaBase):
 class Crawler(CrawlerBase):
     history_capable_date = '2009-01-01'
     schedule = 'Tu,Th'
-    time_zone = -5 
+    time_zone = -5
 
     def crawl(self, pub_date):
         page_url = 'http://www.krakowstudios.com/archive.php?date=%s' % \