+++ comics/comics/pidjin.py
@@ -10,7 +10,6 @@ class Meta(MetaBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 32
-    schedule = 'Mo,Tu,We,Th,Fr,Sa,Su'
     time_zone = -8
 
     def crawl(self, pub_date):