+++ comics/comics/wondermark.py
@@ -11,16 +11,14 @@ class Meta(MetaBase):
 class Crawler(CrawlerBase):
     history_capable_days = 28
     schedule = 'Tu,Fr'
-    time_zone = -8 
+    time_zone = -8
 
     def crawl(self, pub_date):
         feed_url = 'http://feeds.feedburner.com/wondermark'
         feed = self.parse_feed(feed_url)
         for entry in feed.for_date(pub_date):
-            url=entry.content0.src(r'img[src*="/c/"]')
-            title=entry.title
-            text=entry.content0.alt(r'img[src*="/c/"]')
-            # Some of these content blocks are blog posts, not just comics, and
-            # David overlaps them with posts.  Argh.
+            url = entry.content0.src('img[src*="/c/"]')
+            title = entry.title
+            text = entry.content0.alt('img[src*="/c/"]')
             if url is not None:
-                return CrawlerImage(url, title=title, text=text)
+                return CrawlerImage(url, title, text)